Doxygen comment updated.

+++ sensors/adxl345/adxl345.c
@@ -20,7 +20,7 @@ digitalPin aSlaveL = {&TRISBbits, &PORTBbits, 9};
  *         -1 - if device id read is different thed default (factory) id
  *         -2 - if device enabling failed
  */
-UINT8 adxl345Init(digitalPin csPin) {
+int adxl345Init(digitalPin csPin) {
 
      // check if spi2 is enabled
     if (!SPI2STATbits.SPIEN) {
