+++ sensors/adxl345/adxl345.h
@@ -1,5 +1,5 @@
 /*! \file   adxl345.h
- * Declaration of functions and variables used for SPI communication with ADXL345 digital accelerometer sensor.
+ * \brief Declaration of functions and variables used for SPI communication with ADXL345 digital accelerometer sensor.
  */
 
 #ifndef ADXL345_H
@@ -27,9 +27,9 @@
 /*! \brief Macro for ADXL345 command format.
  *
  * Example of usage: \n
- *      reading single byte from reg 0 :                ADXL_COMMAND(0, ADXL_READ) \n
- *      reading multiple bytes/reg starting from reg 0  ADXL_COMMAND(0, ADXL_READ | ADXL_MB) \n
- *      writing a byte to register  0                   ADXL_COMMAND(0, ADXL_WRITE)
+ *      reading single byte from reg 0:                ADXL_COMMAND(0, ADXL_READ) \n
+ *      reading multiple bytes/reg starting from reg 0:  ADXL_COMMAND(0, ADXL_READ | ADXL_MB) \n
+ *      writing a byte to register 0:                   ADXL_COMMAND(0, ADXL_WRITE)
  */
 #define ADXL_COMMAND(reg, options)  ((reg | options) << 8)
 
@@ -56,7 +56,7 @@ enum ADXL_REGISTERS {
     REG_POWER_CTL,          /*!< Power saving feature control register. */
     REG_INT_ENABLE,         /*!< Interrupt enable control register. */
     REG_INT_MAP,            /*!< Interrupt mapping control register. */
-    REG_INT_SOURCE,         /*!< ?ource of interrupts register. */
+    REG_INT_SOURCE,         /*!< Source of interrupts register. */
     REG_DATA_FORMAT,        /*!< Data format control register. */
     REG_DATAX0,             /*!< X axis data the least significant byte register. */
     REG_DATAX1,             /*!< X axis data the most significant byte register. */
@@ -69,16 +69,16 @@ enum ADXL_REGISTERS {
 };
 
 /* Digital pins used as chip select pins for SPI comm*/
-/*! Digital pin used as a SPI chip select pin for the front accelerometer sensor (front = front CASU side).
+/*! Digital pin used as a SPI chip select pin for the front accelerometer sensor.
  */
 extern digitalPin aSlaveF;
-/*! Digital pin used as a SPI chip select pin for the right accelerometer sensor (front = front CASU side).
+/*! Digital pin used as a SPI chip select pin for the right accelerometer sensor.
  */
 extern digitalPin aSlaveR;
-/*! Digital pin used as a SPI chip select pin for the back accelerometer sensor (front = front CASU side).
+/*! Digital pin used as a SPI chip select pin for the back accelerometer sensor.
  */
 extern digitalPin aSlaveB;
-/*! Digital pin used as a SPI chip select pin for the left accelerometer sensor (front = front CASU side).
+/*! Digital pin used as a SPI chip select pin for the left accelerometer sensor.
  */
 extern digitalPin aSlaveL;
 
@@ -89,11 +89,11 @@ extern digitalPin aSlaveL;
  *         -2 - Error in enabling device. \n
  *           1 - Initialization successfully completed.
  */
-UINT8 adxl345Init(digitalPin csPin);
+int adxl345Init(digitalPin csPin);
 
 /*! \brief Function reads accelerometer X axis value.
  * @param  csPin Digital pin used as a SPI chip select pin.
- * @param  ax Memory location where data is stored.
+ * @param  ax Memory location for storing the data.
  * @return 1 - Communication succeeded \n
  *         0 - Communication failed
  */
@@ -101,7 +101,7 @@ UINT8 readAccX(digitalPin csPin, int *ax);
 
 /*! \brief Function reads accelerometer Y axis value.
  * @param  csPin Digital pin used as a SPI chip select pin.
- * @param  ay Memory location where data is stored.
+ * @param  ay Memory location for storing the data.
  * @return 1 - Communication succeeded \n
  *         0 - Communication failed
  */
@@ -109,7 +109,7 @@ UINT8 readAccY(digitalPin csPin, int *ay);
 
 /*! \brief Function reads accelerometer Z axis value.
  * @param  csPin Digital pin used as a SPI chip select pin.
- * @param  ax Memory location where data is stored.
+ * @param  ax Memory location for storing the data.
  * @return 1 - Communication succeeded \n
  *         0 - Communication failed
  */
@@ -117,7 +117,7 @@ UINT8 readAccZ(digitalPin csPin, int *az);
 
 /*! \brief Function reads accelerometer X,Y,Z axis values in a single SPI transmission.
  * @param  csPin Digital pin used as a SPI chip select pin.
- * @param  acc Memory location where data is stored.
+ * @param  acc Memory location for storing the data.
  * @return 1 - Communication succeeded \n
  *         0 - Communication failed
  */
@@ -125,7 +125,7 @@ UINT8 readAccXYZ(digitalPin csPin, int *acc);
 
 /*! \brief Function reads device id.
  * @param  csPin Digital pin used as a SPI chip select pin.
- * @param  id  Memory location where read id is saved.
+ * @param  id  Memory location for storing the data.
  * @return  1 - Communication succeeded. \n
  *          0 - Communication failed.
  */
@@ -134,7 +134,7 @@ UINT8 adxl345readID(digitalPin csPin, UINT16 *id);
 /*! \brief Function reads a single device register.
  * @param  csPin Digital pin used as a SPI chip select pin.
  * @param  reg Address of the register.
- * @param data Memory location where data is stored.
+ * @param data Memory location for storing the data.
  * @return 1 - Communication succeeded.
  *         0 - Communication failed.
  */
